Prava pouzivatelov

+++ zapis_poruch/linka/views.py
@@ -9,29 +9,29 @@ from django.http import HttpResponse
 from django.shortcuts import render, redirect
 from django.views.generic import View
 from .forms import TypForm, ZaznamForm, RevizieForm
-from linka.management.commands.seed import run_seed
-from django.contrib.auth.models import User
 from .models import TypChyby, Chyba, TypRevizie, ChybaWrapper, TypChybyWrapper, DruhChyby, \
     MiestoNaLinke, SposobenaKym
-
-
-class Seed(View):
-    def get(self, request):
-        run_seed("")
-        next = request.POST.get('next', '/')
-        return redirect(next)
 
 
 class TypyChyb(LoginRequiredMixin, View):
     template = "chyby_typy.html"
 
     def get(self, request):
+        group_permissions = request.user.get_group_permissions()
+
+        if 'linka.view_typchyby' not in group_permissions:
+            print('Prístup odmietnutý')
+            return render(request, 'pristup_zakazany.html', {})
+
+
         all_errors = ChybaWrapper.all()
         all_types = TypChybyWrapper.all()
         for object in all_types:
             object.fill(all_errors)
 
-        data = {'errors': [x.json() for x in all_types]}
+        data = {'errors': [x.json() for x in all_types],
+                'permissions': group_permissions
+                }
 
         if "order_by" in request.GET:
             order_by = request.GET.get('order_by', 'defaultOrderField')
@@ -57,12 +57,20 @@ class Zaznamy(LoginRequiredMixin, View):
     template = "zaznamy.html"
 
     def get(self, request):
+        group_permissions = request.user.get_group_permissions()
+
+        if 'linka.view_chyba' not in group_permissions:
+            print('Prístup odmietnutý')
+            return render(request, 'pristup_zakazany.html', {})
+
         if "delete" in request.GET:
             i = request.GET["id"]
             chyba = Chyba.objects.all().filter(id=i)
             chyba.delete()
 
-        data = {'zaznamy': ChybaWrapper.all()}
+        data = {'zaznamy': ChybaWrapper.all(),
+                'permissions': group_permissions
+                }
 
         if "order_by" in request.GET:
             order_by = request.GET.get('order_by', 'defaultOrderField')
@@ -85,6 +93,8 @@ class Zaznamy(LoginRequiredMixin, View):
                 data['zaznamy'] = sorted(data['zaznamy'], key=lambda obj: obj.dovod)
             if order_by == "opatrenie":
                 data['zaznamy'] = sorted(data['zaznamy'], key=lambda obj: obj.opatrenia)
+        else:
+            data['zaznamy'] = sorted(data['zaznamy'], key=lambda obj: (obj.schvalena, obj.vyriesena))
 
         return render(request, self.template, data)
 
@@ -96,6 +106,12 @@ class PridajTyp(LoginRequiredMixin, View):
     template = "pridaj_typ.html"
 
     def get(self, request):
+        group_permissions = request.user.get_group_permissions()
+
+        if 'linka.add_typchyby' not in group_permissions and 'linka.change_typchyby' not in group_permissions:
+            print('Prístup odmietnutý')
+            return render(request, 'pristup_zakazany.html', {})
+
         data = dict()
 
         if "id" not in request.GET:
@@ -107,6 +123,12 @@ class PridajTyp(LoginRequiredMixin, View):
         return render(request, self.template, data)
 
     def post(self, request):
+        group_permissions = request.user.get_group_permissions()
+
+        if 'linka.add_typchyby' not in group_permissions and 'linka.change_typchyby' not in group_permissions:
+            print('Prístup odmietnutý')
+            return render(request, 'pristup_zakazany.html', {})
+
         if "id" in request.GET:
             typ = TypChyby.objects.all().filter(id=request.GET["id"])[0]
             form = TypForm(request.POST, instance=typ)
@@ -123,6 +145,12 @@ class PridajZaznam(LoginRequiredMixin, View):
     template = "pridaj_zaznam.html"
 
     def get(self, request):
+        group_permissions = request.user.get_group_permissions()
+
+        if 'linka.add_chyba' not in group_permissions and 'linka.change_chyba' not in group_permissions:
+            print('Prístup odmietnutý')
+            return render(request, 'pristup_zakazany.html', {})
+
         data = dict()
 
         if "id" not in request.GET:
@@ -137,6 +165,12 @@ class PridajZaznam(LoginRequiredMixin, View):
         return render(request, self.template, data)
 
     def post(self, request):
+        group_permissions = request.user.get_group_permissions()
+
+        if 'linka.add_chyba' not in group_permissions and 'linka.change_chyba' not in group_permissions:
+            print('Prístup odmietnutý')
+            return render(request, 'pristup_zakazany.html', {})
+
         if "id" in request.GET:
             typ = Chyba.objects.all().filter(id=request.GET["id"])[0]
             form = ZaznamForm(request.POST, instance=typ)
@@ -153,6 +187,12 @@ class PridajRevizia(LoginRequiredMixin, View):
     template = "pridaj_revizia.html"
 
     def get(self, request):
+        group_permissions = request.user.get_group_permissions()
+
+        if 'linka.add_typrevizie' not in group_permissions and 'linka.change_typrevizie' not in group_permissions:
+            print('Prístup odmietnutý')
+            return render(request, 'pristup_zakazany.html', {})
+
         data = dict()
 
         if "id" not in request.GET:
@@ -164,6 +204,12 @@ class PridajRevizia(LoginRequiredMixin, View):
         return render(request, self.template, data)
 
     def post(self, request):
+        group_permissions = request.user.get_group_permissions()
+
+        if 'linka.add_typrevizie' not in group_permissions and 'linka.change_typrevizie' not in group_permissions:
+            print('Prístup odmietnutý')
+            return render(request, 'pristup_zakazany.html', {})
+
         if "id" in request.GET:
             typ = TypRevizie.objects.all().filter(id=request.GET["id"])[0]
             form = RevizieForm(request.POST, instance=typ)
@@ -180,6 +226,11 @@ class Revizia(LoginRequiredMixin, View):
     template = "revizia.html"
 
     def get(self, request):
+        group_permissions = request.user.get_group_permissions()
+
+        if 'linka.view_typrevizie' not in group_permissions:
+            print('Prístup odmietnutý')
+            return render(request, 'pristup_zakazany.html', {})
 
         if "delete" in request.GET:
             i = request.GET["id"]
@@ -193,7 +244,11 @@ class Revizia(LoginRequiredMixin, View):
             revizia.datum_nadchadzajucej_revizie = date.today() + timedelta(days=int(revizia.exspiracia))
             revizia.save()
 
-        data = {'revizie': TypRevizie.objects.all(), 'today': date.today(), 'weeks': date.today() + timedelta(days=28)}
+        data = {'revizie': TypRevizie.objects.all(),
+                'today': date.today(),
+                'weeks': date.today() + timedelta(days=28),
+                'permissions': group_permissions
+                }
 
         if "order_by" in request.GET:
             order_by = request.GET.get('order_by', 'defaultOrderField')
@@ -218,15 +273,29 @@ class Grafy(LoginRequiredMixin, View):
     template = "grafy.html"
 
     def get(self, request):
+        group_permissions = request.user.get_group_permissions()
+
+        if 'linka.view_grafy' not in group_permissions:
+            print('Prístup odmietnutý')
+            return render(request, 'pristup_zakazany.html', {})
+
         data = {
             "druhyChyb": DruhChyby.objects.all(),
             "zariadenia": MiestoNaLinke.objects.all(),
             "sposobeneKym": SposobenaKym.objects.all(),
-            "popisyTypovChyby": TypChyby.objects.all()
+            "popisyTypovChyby": TypChyby.objects.all(),
+            "permissions": group_permissions,
         }
+
         return render(request, self.template, data)
 
     def post(self, request):
+        group_permissions = request.user.get_group_permissions()
+
+        if 'linka.view_grafy' not in group_permissions:
+            print('Prístup odmietnutý')
+            return render(request, 'pristup_zakazany.html', {})
+
         def getInt(val):
             try:
                 return int(val)
@@ -282,19 +351,27 @@ class Grafy(LoginRequiredMixin, View):
             "popisyTypovChyby": TypChyby.objects.all(),
             "grafLabels": grafLabels,
             "grafColors": grafColors,
-            "grafData": grafData
+            "grafData": grafData,
+            "permissions": group_permissions
         }
         return render(request, self.template, data)
 
 
-class Email(View):
+class Email(LoginRequiredMixin, View):
     template = "email.html"
 
     def get(self, request):
+        if not request.user.is_superuser:
+            print('Prístup odmietnutý')
+            return render(request, 'pristup_zakazany.html', {})
 
         return render(request, self.template, {})
 
     def post(self, request):
+        if not request.user.is_superuser:
+            print('Prístup odmietnutý')
+            return render(request, 'pristup_zakazany.html', {})
+
         mail_list = ['namova9094@pyrelle.com']  # , 'freyer.viktor@gmail.com']
         now = datetime.datetime.now()
 
@@ -342,10 +419,16 @@ class Email(View):
         return redirect('email')
 
 
-class PotvrdZaznam(View):
+class PotvrdZaznam(LoginRequiredMixin, View):
     template = "potvrd_zaznam.html"
 
     def get(self, request):
+        group_permissions = request.user.get_group_permissions()
+
+        if 'linka.approve_chyba' not in group_permissions:
+            print('Prístup odmietnutý')
+            return render(request, 'pristup_zakazany.html', {})
+
         if "put" in request.GET:
             i = request.GET["id"]
             zaznam = Chyba.objects.all().filter(id=i)[0]
@@ -361,6 +444,12 @@ class PotvrdZaznam(View):
             return render(request, self.template, data)
 
     def post(self, request):
+        group_permissions = request.user.get_group_permissions()
+
+        if 'linka.approve_chyba' not in group_permissions:
+            print('Prístup odmietnutý')
+            return render(request, 'pristup_zakazany.html', {})
+
         i = request.GET["id"]
         zaznam = Chyba.objects.all().filter(id=i)[0]
         typ = request.GET["list"]
@@ -369,23 +458,13 @@ class PotvrdZaznam(View):
         return redirect("zaznamy")
 
 
-class Pouzivatelia(LoginRequiredMixin, View):
-    template = "pouzivatelia.html"
-
-    def get(self, request):
-        if "delete" in request.GET and request.GET["delete"]:
-            i = request.GET["id"]
-            pouzivatel = User.objects.all().filter(id=i)
-            pouzivatel.delete()
-        data = {"pouzivatelia": User.objects.all()}
-        return render(request, self.template, data)
-
-    def post(self, request):
-        return HttpResponse('podarilo sa')
-
-
 class Login(LoginView):
     template_name = "login.html"
+
+    def get(self, request, *args, **kwargs):
+        if request.user.is_authenticated:
+            return redirect("/")
+        return super().get(request, args, kwargs)
 
 
 class Logout(View):
